Update to handle complex triggers
